requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.9.4
amqp==5.1.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
autocommand==2.2.2
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backoff==2.2.1
backports.csv==1.0.7
beautifulsoup4==4.11.2
bentoml==1.0.15
billiard==3.6.4.0
blis==0.7.9
boto3==1.24.59
botocore==1.27.59
build==0.10.0
cachetools==5.3.0
catalogue==2.0.8
cattrs==22.2.0
celery==5.2.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
cheroot==9.0.0
CherryPy==18.8.0
circus==0.18.0
click==8.1.3
click-didyoumean==0.3.0
click-option-group==0.5.5
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm @ file:///croot/comm_1671231121260/work
confection==0.0.4
configobj==5.0.8
contextlib2==21.6.0
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
cymem==2.0.7
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.4
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepmerge==1.1.0
Deprecated==1.2.13
-e git+https://github.com/DTS-Product/MLOps-DiabetesPrediction.git@0913e1633593dace57006fcb6bc05f5b77d897eb#egg=DiabetesPredictionApp
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.4.0
distlib==0.3.6
distro==1.8.0
docker==6.0.1
dpath==2.1.4
dulwich==0.21.3
dvc==2.45.1
dvc-data==0.40.3
dvc-gdrive==2.19.1
dvc-http==2.30.2
dvc-objects==0.19.3
dvc-render==0.2.0
dvc-s3==2.21.0
dvc-studio-client==0.4.0
dvc-task==0.1.11
dvclive==2.1.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
exceptiongroup==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
feedparser==6.0.10
filelock==3.9.0
Flask==2.2.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.1.0
funcy==1.18
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
grandalf==0.8
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httplib2==0.21.0
huggingface-hub==0.12.1
hydra-core==1.3.1
idna==3.4
imageio==2.25.1
importlib-metadata==5.2.0
importlib-resources==5.12.0
inflect==6.0.2
iniconfig==2.0.0
ipykernel @ file:///croot/ipykernel_1671488378391/work
ipython @ file:///croot/ipython_1676582224036/work
iterative-telemetry==0.0.7
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter-dash==0.4.2
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1676538566912/work
kaleido==0.2.1
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
lime==0.2.0.1
llvmlite==0.39.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mlflow==2.1.1
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.14
munch==2.5.0
murmurhash==1.0.9
mysqlclient==2.1.1
nanotime==0.5.2
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==3.0
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
omnixai==1.2.5
opencv-python-headless==4.7.0.68
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
orjson==3.8.6
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.5.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.11.0
pathy==0.10.1
patsy==0.5.3
Pattern==3.6
pdfminer.six==20221105
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pip-requirements-parser==32.0.1
pip-tools==6.12.2
platformdirs==3.0.0
plotly==5.13.0
pluggy==1.0.0
polyjuice-nlp==0.1.5
portend==3.1.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pydot==1.4.2
PyDrive2==1.15.0
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
PyJWT==2.6.0
pynvml==11.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproject_api==1.5.0
pyproject_hooks==1.0.0
pytest==7.2.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-docx==0.8.11
python-json-logger==2.0.6
python-multipart==0.0.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.2
retrying==1.3.4
rich==13.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2023.1.0
s3transfer==0.6.0
SALib==1.4.7
schema==0.7.5
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.0
scmrepo==0.1.11
sentence-transformers==2.2.2
sentencepiece==0.1.97
sgmllib3k==1.0.0
shap==0.41.0
shortuuid==1.0.11
shtab==1.5.8
simple-di==0.1.5
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.46
sqlparse==0.4.3
sqltrie==0.0.28
srsly==2.4.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.25.0
statsmodels==0.13.5
tabulate==0.9.0
tempora==5.2.1
tenacity==8.2.1
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2023.2.3
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
torchvision==0.14.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tox==4.2.6
tqdm==4.64.1
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
virtualenv==20.19.0
voluptuous==0.13.1
wasabi==1.1.1
watchfiles==0.18.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zc.lockfile==2.0
zipp==3.14.0
zss==1.2.0
